Master 1.12 dev (#169)
* Cleanups and fixes for dependencies, block and item registration for
the 1.12 port of BaseMetals. All that is left, at this point, is to port
the recipes that we can to the JSON system and fix the others to work
using the recipe registry

This also changes the code to disable the Achievements until they can be
ported as well.

* More disabling of achievements code and the first update to shift the
block&item registration over to the new methods introduced with 1.12

These have to be changed to work on a per-mod basis, however.

* Fix things up so the code registers only those materials created by
the current, active mod so we work within the restrictions of how the
new registry system in 1.12 is designed.

This should be the last before the recipe changes.

* More porting work - this fixes the villager trades, further breaks the shield upgrades
on the anvil, removes (hopefully) the last vestiges of the old achievements code and
gets things ready for the conditions to get added for the json recipe system.

I think we still have some work to do somewhere else, but at this time the only errors
are in the unported recipe code and the Anvil Event stuff for the shield upgrades.

* Last of the fixes and cleanups not related to the system actually getting the
JSON recipes. I'm hoping that the code for the Shield Upgrades, especially the
way I've mangled it, still works but if not, I have other ideas for it.

The Repair Recipes need a factory, the alloys need a condition that checks for
the existence of the source materials and everything needs the condition for
things being able to be turned off in config.

Thats right, we're down to only 4 things having to be done!

* Update to 1.12.1 (need to update the mappings still), comment out some unused deps
and generally update to newer versions of a few of the other deps.

* add a simple tool for generating the recipe jsons - this should come in handy for other mods as well

* add the conditions and recipe factories needed to cover our uses here

* this now does about 66% of its job - just needs to have the vanilla recipes output added and start writing to disk

* finish the tool for generating the recipes and move the generated recipes into the assets directory and the (hopefully) correct place in it

* remove most of the outdated recipe registration code that is no longer needed

* fix a thinko in the generator - it wasn't writing out the _factories.json - and add the _factories.json to the repo

* add a new condition to cover for repair-armor-with-plate recipes and a new factory to cover for the actual armor repair recipes

* I made a mistake in setting up the ingredient and result stuff originally, leading to it using actual ore-dictionary names instead of using the 'forge:ore_dict' ingredient factory

* final fixes to bring the armor repair (hopefully) online, this should be the last of the changes to the recipe lineup and this form of the generating tool

* Final set of changes and fixes that get this compiling against Forge for MC 1.12.1

1) The churn in the recipes is because of a re-gen after I accidentally deleted a few
   and didn't want them showing as deleted.
1a) Still managed to commit a delete
2) Final 10 "these are enhancements or extensions of vanilla recipes" - all relating to
   steel have been hand-ported to the JSON format
3) Last chunks of cleanups and removal of unused code re: port have been done
4) Buckets of Mercury are still handled in the code instead of in a JSON because that
   one recipe would have required a custom factory.
5) More cleanups will probably follow as I find emacs backups and crash-proofing files

* found a backup file

* Testing turned up a logic bug in the handling of input ingredients for the
shapeless recipes. A more proper fix would be to require the ingredients to
always be an array.

For now, though, this fixes the use-cases

* another logic bug in the auto-gen code

* actual testing turned up quite a few problems with the recipes and some thinko's elsewhere in the code - this is an intermediate step towards getting that all resolved

* more logic fixes

* logic errors that were halting the game from loading are solved, still have to solve the issues with the recipes and stuff not, apparently, getting registered