Finish up getting Tinkers' working correctly.
During the update process from 1.10.2 to the 1.12 series we seem to have lost, along the way, most of the casting of goods in Tinkers' - this patch fixes that. Now all materials that support the various items can have them cast, though the Tinkers' "tool rod" takes precedence over the rods we provide and I haven't been able to get Diamond and Emerald to do exactly what I'd like - but that should come at some point as I get more used to the code.

If I was actually going to go whole-hog on this I'd have to flip a few switches elsewhere in the code and do some big-time manual work in the same manner that the code in Tinkers' does for registering the casting recipes of its own stuff.