Seems the events, because they are being fired during the overall FML events, had some issues as things were originally done. Even with the change to having a bit of metadata tracking there were issues with "dangerous alternative prefix" messages being spewed out.
Use the public methods of FML and the metadata tracking already in place to swap around the active mod container and clear those issues. Next step is back to ModernMetals to see what is halting the TiCon integration from working correctly. This issue might, actually, be related to some simple timing issues.