ItemMMDSickle should derive from net.minecraft.item.ItemTool instead of com.mcmoddev.lib.item.GenericMMDItem - implement this change. (#344)
ItemMMDSickle had getItemAttributeModifiers as an @Override, but was calling super.getAttributeModifiers(), leading to a circular call-chain, stack overflow and crash. Fix this by having it call the proper super method.
GenericMMDItem had some support for generic tools added during localising the bug - remove it and do a tiny bit of whitespace cleanup that was needed because of that and some other tiny bits.
GenericMMDItem has gained an import of net.minecraft.item.Item to make the class declaration a bit cleaner
com.mcmoddev.basemetals.BaseMetals had a slew of unused imports. Those have been removed.